% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BSplines.R
\name{BSplines}
\alias{BSplines}
\title{builds B-Splines Matrix with appropriate knots for taper fitting}
\usage{
BSplines(knots = c(seq(0, 1, 0.1)), ord = 4, der = 0, x = c(seq(0, 1, 0.01)), ...)
}
\arguments{
\item{knots}{knot positions for spline function}

\item{ord}{order of the spline function}

\item{der}{derivatives}

\item{x}{height measurements}

\item{...}{not currently used}
}
\value{
B-Splines matrix build using \code{\link[splines]{splineDesign}}
}
\description{
Internal function not usually called by users
}
\details{
internally \code{\link[splines]{splineDesign}} is called
}
\author{
Edgar Kublin
}
