% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dN.f.R
\name{dN.f}
\alias{dN.f}
\title{evaluate Normal distribution}
\usage{
dN.f(x, mw, sd, ...)
}
\arguments{
\item{x}{vector of quantiles}

\item{mw}{vector of means}

\item{sd}{vector of standard deviations}

\item{...}{not currently used}
}
\value{
numeric density of normal distribution
}
\description{
Internal function not usually called by users
}
\author{
Edgar Kublin
}
