% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EYx_ssp.f.R
\name{EYx_ssp.f}
\alias{EYx_ssp.f}
\title{estimate cubic smoothing spline with linear extrapolation}
\usage{
EYx_ssp.f(knt, coe, x, ...)
}
\arguments{
\item{knt}{knots position of B-Splines}

\item{coe}{estimated coefficient for B-Splines}

\item{x}{position at which to evaluate B-Splines model}

\item{...}{not currently used}
}
\value{
expected diameter given knots and coefficients at position (height)
\code{x}.
}
\description{
Internal function not usually called by users
}
\author{
Edgar Kublin
}
