% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TransKnots.R
\name{TransKnots}
\alias{TransKnots}
\title{transform knot vector}
\usage{
TransKnots(knots = c(seq(0, 1, 0.1)), ord = 4, ...)
}
\arguments{
\item{knots}{knot positions for spline function}

\item{ord}{order of the spline function}

\item{...}{not currently used}
}
\value{
transformed knots vector, especially with repeated first and last 
knot given order of spline function
}
\description{
Internal function not usually called by users
}
\author{
Edgar Kublin
}
