% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/updateKOV.R
\name{updateKOV}
\alias{updateKOV}
\title{update variance-covariance matrix}
\usage{
updateKOV(var, kov)
}
\arguments{
\item{var}{vector of new variances}

\item{kov}{variance-covariance matrix to be updated}
}
\value{
updated variance-covariance matrix
}
\description{
Internal function not usually called by users
}
\details{
Firstly, \code{kov} is transformed to a correlation matrix. 
Secondly, this correlation matrix is backtransformed to a variance-covariance
matrix using the new variances.
}
\author{
Christian Vonderach
}
