% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xy0_root.f.R
\name{xy0_root.f}
\alias{xy0_root.f}
\title{xy0_root.f}
\usage{
xy0_root.f(x, y0, SK, par.lme, ...)
}
\arguments{
\item{x}{relative height}

\item{y0}{diameter for which height is required}

\item{SK}{return of \code{\link{SK_EBLUP_LME.f}} containing estimated fixed
and the tree specific random effects of the taper model}

\item{par.lme}{List of taper model parameters obtained by \code{\link{TapeR_FIT_LME.f}}}

\item{...}{not currently used}
}
\value{
difference between actual diameter at height \code{x} and given
diameter \code{y0}
}
\description{
Internal function not usually called by users
}
\details{
used in \code{xy0_SK_EBLUP_LME.f} to find the root of taper curve
(i.e. height \code{x}) at given diameter \code{y0}
}
\author{
Edgar Kublin
}
