% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xy0_SK_EBLUP_LME.f.R
\name{xy0_SK_EBLUP_LME.f}
\alias{xy0_SK_EBLUP_LME.f}
\title{find relative height of given diameter}
\usage{
xy0_SK_EBLUP_LME.f(xm, ym, y0, par.lme, Rfn = list(fn = "sig2"), ...)
}
\arguments{
\item{xm}{relative heights for which measurements are available}

\item{ym}{corresponding diameter measurements in height \code{xm}}

\item{y0}{given diameter for which height is required}

\item{par.lme}{Fitted model object, return of \code{\link{TapeR_FIT_LME.f}}}

\item{Rfn}{list with function name to provide estimated or assumed residual 
variances for the given measurements, optionally parameters for such functions}

\item{...}{not currently used}
}
\value{
relative height of given diameter \code{y0}
}
\description{
Internal function not usually called by users
}
\details{
function used to transform given diameter in volume calculation into
height; c.f \code{\link{E_VOL_AB_HmDm_HT.f}}; with \code{Rfn} one can decide 
whether the measured diameters are forced to lie exactly on the taper curve
\code{Rfn$fn="sig2"} or not \code{Rfn$fn="zero"}. Other options are possible, 
see also \code{\link{SK_EBLUP_LME.f}} and \code{\link{resVar}}.
}
\author{
Edgar Kublin
}
