% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/XZ_BSPLINE.f.R
\name{XZ_BSPLINE.f}
\alias{XZ_BSPLINE.f}
\title{construct B-Splines design matrices}
\usage{
XZ_BSPLINE.f(x, knt, ord, ...)
}
\arguments{
\item{x}{relative height measurements}

\item{knt}{knot positions for B-Splines, usually taken from a model fit by 
\code{\link{TapeR_FIT_LME.f}}}

\item{ord}{order of B-Splines, usually taken from a model fit by 
\code{\link{TapeR_FIT_LME.f}}}

\item{...}{not currently used}
}
\value{
List with height measurements (\code{x}), the fixed effects B-splines
matrix and the random effects B-splines matrix.
}
\description{
Internal function not usually called by users
}
\seealso{
\code{\link{TapeR_FIT_LME.f}}
}
\author{
Edgar Kublin
}
