% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/y2x_isp.f.R
\name{y2x_isp.f}
\alias{y2x_isp.f}
\title{squared diameter using interpolating splines}
\usage{
y2x_isp.f(x, x.grd, y.grd, ...)
}
\arguments{
\item{x}{relative height}

\item{x.grd}{relative heights for interpolation}

\item{y.grd}{diameter of taper curve at relative heights \code{x.grd} for
interpolation}

\item{...}{not currently used}
}
\value{
squared estimated diameter based on natural interpolating spline 
(\code{\link[stats]{splinefun}})
}
\description{
Internal function not usually called by users
}
\author{
Edgar Kublin
}
