% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yx_ssp.f.R
\name{yx_ssp.f}
\alias{yx_ssp.f}
\title{diameter using smoothing splines}
\usage{
yx_ssp.f(x, x.grd, y.grd, ...)
}
\arguments{
\item{x}{relative height}

\item{x.grd}{relative heights for interpolation}

\item{y.grd}{diameter of taper curve at relative heights \code{x.grd} for
interpolation}

\item{...}{not currently used}
}
\value{
estimated diameter based on smoothing splines 
(\code{\link[stats]{smooth.spline}})
}
\description{
Internal function not usually called by users
}
\author{
Edgar Kublin
}
