% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SK_VOLab_EBLUP_LME.f.R
\name{SK_VOLab_EBLUP_LME.f}
\alias{SK_VOLab_EBLUP_LME.f}
\title{taper volume estimation}
\usage{
SK_VOLab_EBLUP_LME.f(xm, ym, a = 0, b = 1, Ht, par.lme, IntPolOpt = T, ...)
}
\arguments{
\item{xm}{relative heights for which measurements are available}

\item{ym}{corresponding diameter measurements in height \code{xm}}

\item{a}{relative height of lower threshold of stem section}

\item{b}{relative height of upper threshold of stem section}

\item{Ht}{tree height}

\item{par.lme}{List of taper model parameters obtained by \code{\link{TapeR_FIT_LME.f}}}

\item{IntPolOpt}{option for method of interpolation, if TRUE using a natural 
interpolating spline (\code{\link[stats]{splinefun}}), if FALSE using a smoothing 
spline (\code{\link[stats]{smooth.spline}}); defaults to TRUE}

\item{...}{not currently used}

\item{A}{Numeric scalar defining the lower threshold of a stem section for volume
estimation. Depends on \code{iDH}. If \code{iDH} = "D", a diameter
(cm), if \code{iDH} = "H", a height (m). If NULL, section starts at lowest point.}

\item{B}{Numeric scalar defining the upper threshold of a stem section for volume
estimation. Depends on \code{iDH}. If \code{iDH} = "D", a diameter
(cm), if \code{iDH} = "H", a height (m). If NULL, section ends at tip.}
}
\value{
List with two elements, the estimated volume and its variance
}
\description{
Internal function not usually called by users
}
\author{
Edgar Kublin
}
